+++ e2e/tests/rmqTest.e2e-spec.ts
@@ -1,24 +1,25 @@
 import { Test } from '@nestjs/testing';
 import { RMQModule, RMQService } from '../../lib';
 import { INestApplication } from '@nestjs/common';
-import { ApiController } from '../mocks/api.controller';
 import { MicroserviceController } from '../mocks/microservice.controller';
-import { ConfigModule } from '../mocks/config.module';
-import { ConfigService } from '../mocks/config.service';
 import { IRMQService } from '../../lib/interfaces/rmq-service.interface';
-import { DivideContracts } from '../contracts/mock.contracts';
+import { AppIdContracts, CustomMessageFactoryContracts, DivideContracts, MultiplyContracts, PatternHashContracts, PatternStarContracts, SumContracts } from '../contracts/mock.contracts';
+import { ERROR_UNDEFINED_FROM_RPC } from '../../lib/constants';
+import { DoublePipe } from '../mocks/double.pipe';
+import { ZeroIntercepter } from '../mocks/zero.intercepter';
 
-describe('RMQe2e', () => {
+describe('RMQe2e forTest()', () => {
 	let api: INestApplication;
-	let apiController: ApiController;
-	let microserviceController: MicroserviceController;
 	let rmqService: IRMQService;
 
 	beforeAll(async () => {
 		const apiModule = await Test.createTestingModule({
 			imports: [
-				ConfigModule,
-				RMQModule.forTest({})
+				RMQModule.forTest({
+					serviceName: 'test-service',
+					middleware: [DoublePipe],
+					intercepters: [ZeroIntercepter],
+				})
 			],
 			controllers: [MicroserviceController],
 		}).compile();
@@ -29,12 +30,129 @@ describe('RMQe2e', () => {
 	});
 
 	describe('Running methods', () => {
-		it('triggerRoute', async () => {
-			const res = rmqService.triggerRoute<DivideContracts.Request, DivideContracts.Response>(DivideContracts.topic, {
+		it('successful send()', async () => {
+			const { result } = await rmqService.triggerRoute<SumContracts.Request, SumContracts.Response>(SumContracts.topic, {
+				arrayToSum: [1, 2, 3]
+			});
+			expect(result).toEqual(6);
+		});
+		it('successful appId from message', async () => {
+			const { appId } = await rmqService.triggerRoute<null, AppIdContracts.Response>(AppIdContracts.topic, null);
+			expect(appId).toBe('test-service');
+		});
+		it('request validation failed', async () => {
+			try {
+				await rmqService.triggerRoute<any, SumContracts.Response>(SumContracts.topic, {
+					arrayToSum: ['a', 'b', 'c']
+				});
+				expect(true).toBe(false);
+			} catch (error) {
+				expect(error.message).toBe(
+					'each value in arrayToSum must be a number conforming to the specified constraints',
+				);
+			}
+		});
+		it('get common Error from method', async () => {
+			try {
+				const { result } = await rmqService.triggerRoute<any, SumContracts.Response>(SumContracts.topic, {
+					arrayToSum: [0, 0, 0]
+				});
+				expect(result).not.toBe(0);
+			} catch (error) {
+				expect(error.message).toBe('My error from method');
+			}
+		});
+		it('get RMQError from method', async () => {
+			try {
+				const { result } = await rmqService.triggerRoute<any, SumContracts.Response>(SumContracts.topic, {
+					arrayToSum: [-1, 0, 0]
+				});
+				expect(result).not.toBe(-1);
+			} catch (error) {
+				expect(error.message).toBe('My RMQError from method');
+				expect(error.type).toBe('RMQ');
+				expect(error.code).toBe(0);
+				expect(error.data).toBe('data');
+			}
+		});
+		it('get undefined return Error', async () => {
+			try {
+				const { result } = await rmqService.triggerRoute<any, SumContracts.Response>(SumContracts.topic, {
+					arrayToSum: [-11, 0, 0]
+				});
+				expect(result).not.toBe(-11);
+			} catch (error) {
+				expect(error.message).toBe(ERROR_UNDEFINED_FROM_RPC);
+				expect(error.code).toBeUndefined();
+				expect(error.data).toBeUndefined();
+			}
+		});
+	});
+
+	describe('Mock results', () => {
+		it('Mock reply', async () => {
+			const res = { a: 1 };
+			const topic = 'a';
+			rmqService.mockReply(topic, res);
+			const data = await rmqService.send(topic, '');
+			expect(data).toEqual(res);
+		});
+
+		it('Mock error', async () => {
+			const error = new Error('error');
+			const topic = 'a';
+			rmqService.mockError(topic, error);
+			try {
+				const data = await rmqService.send(topic, '');
+				expect(true).toBeFalsy();
+			} catch (e) {
+				expect(e.message).toEqual('error');
+			}
+		});
+	});
+
+	describe('middleware', () => {
+		it('doublePipe', async () => {
+			const { result } = await rmqService.triggerRoute<MultiplyContracts.Request, MultiplyContracts.Response>(MultiplyContracts.topic, {
+				arrayToMultiply: [1, 2]
+			});
+			expect(result).toBe(8);
+		});
+	});
+
+	describe('interceptor', () => {
+		it('zeroInterceptor', async () => {
+			const { result } = await rmqService.triggerRoute<DivideContracts.Request, DivideContracts.Response>(DivideContracts.topic, {
 				first: 10,
-				second: 2
-			}, 1);
-			expect(res).toEqual(5);
+				second: 5
+			});
+			expect(result).toBe(0);
+		});
+	});
+
+	describe('msgFactory', () => {
+		it('customMessageFactory', async () => {
+			const { num, appId } = await rmqService.triggerRoute<CustomMessageFactoryContracts.Request, CustomMessageFactoryContracts.Response>(CustomMessageFactoryContracts.topic, {
+				num: 1
+			});
+			expect(num).toBe(2);
+			expect(appId).toBe('test-service');
+		});
+	});
+
+	describe('msgPattent', () => {
+		it('* pattern', async () => {
+			const { num } = await rmqService.triggerRoute<PatternStarContracts.Request, PatternStarContracts.Response>(PatternStarContracts.topic, {
+				num: 1
+			});
+			expect(num).toBe(1);
+		});
+
+		it('# pattern', async () => {
+			const { num } = await rmqService.triggerRoute<PatternHashContracts.Request, PatternHashContracts.Response>(PatternHashContracts.topic, {
+				num: 1
+			});
+			expect(num).toBe(1);
 		});
 	});