+++ lib/rmq.service.spec.ts
@@ -1,5 +1,3 @@
-import { Test } from '@nestjs/testing';
-import { INestApplication } from '@nestjs/common';
 import { RMQService } from './rmq.service';
 import { RMQMetadataAccessor } from './rmq-metadata.accessor';
 import { Reflector } from '@nestjs/core';